core_service/requirements.txt
asgiref==3.8.1
Django==5.2.3
sqlparse==0.5.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework_simplejwt==5.5.0
djangorestframework-simplejwt[django]
debugpy
psycopg2-binary>=2.9
drf-spectacular
drf-spectacular[sidecar]
celery 
redis
playwright
python_dotenv
flower
PyJWT
pyairtable